requirements.txt
pymongo
pyyaml
certifi
